move all core routines needed for compute_MVBS into the python script mvbs.py
